% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{adverse_event}
\alias{adverse_event}
\title{Return a data frame that records the adverse events and their counts and
percent for 2 treatments.}
\usage{
adverse_event(x, biomarker = c("Wild-type", "Mutant"))
}
\arguments{
\item{x}{a data.frame records patients' adverse events situations.}

\item{biomarker}{Must be one of "Wild-type" and "Mutant". If "Wild-type", the
function will return the adverse event data frame for Wild-type KRAS. If
"Mutant", the function will return the adverse event data frame for Mutant
KRAS.}
}
\value{
a dataframe records adverse events' count and percent for each
treatment and the p-value.
}
\description{
Based on the given data, only adverse events that are statistically
significant different between 2 treatments are preserved. For each adverse
event, their count and percent for each treatment and the p-value is
recorded.
}
