% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_surv.R
\name{ae_surv}
\alias{ae_surv}
\title{Classify each patient's adverse event grade as "less than 2" or
"greater or equal to 2".}
\usage{
ae_surv(x, adverse)
}
\arguments{
\item{x}{an object inherited from a data.frame which records patients'
adverse events situations.}

\item{adverse}{the adverse event that want to be classified on.}
}
\value{
a data frame records patients' adverse event classification, as well
as their baseline characteristics, treatments and survival status.
}
\description{
Based on patients' adverse event grade/severity grade, each of their grade is
classified as "less than 2" or "greater or equal to 2".
}
