% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{\%*\%,bis620_sparse_matrix,matrix-method}
\alias{\%*\%,bis620_sparse_matrix,matrix-method}
\alias{\%*\%,bis620_sparse_matrix,dgeMatrix-method}
\alias{\%*\%,bis620_sparse_matrix,dgCMatrix-method}
\alias{\%*\%,bis620_sparse_matrix,bis620_sparse_matrix-method}
\alias{\%*\%,dgCMatrix,matrix-method}
\alias{\%*\%,dgCMatrix,dgeMatrix-method}
\alias{\%*\%,dgCMatrix,dgCMatrix-method}
\alias{\%*\%,dgCMatrix,bis620_sparse_matrix-method}
\alias{\%*\%,dgeMatrix,matrix-method}
\alias{\%*\%,dgeMatrix,dgeMatrix-method}
\alias{\%*\%,dgeMatrix,dgCMatrix-method}
\alias{\%*\%,dgeMatrix,bis620_sparse_matrix-method}
\title{bis620_matrix_multiply_1}
\usage{
\S4method{\%*\%}{bis620_sparse_matrix,matrix}(x, y)

\S4method{\%*\%}{bis620_sparse_matrix,dgeMatrix}(x, y)

\S4method{\%*\%}{bis620_sparse_matrix,dgCMatrix}(x, y)

\S4method{\%*\%}{bis620_sparse_matrix,bis620_sparse_matrix}(x, y)

\S4method{\%*\%}{dgCMatrix,matrix}(x, y)

\S4method{\%*\%}{dgCMatrix,dgeMatrix}(x, y)

\S4method{\%*\%}{dgCMatrix,dgCMatrix}(x, y)

\S4method{\%*\%}{dgCMatrix,bis620_sparse_matrix}(x, y)

\S4method{\%*\%}{dgeMatrix,matrix}(x, y)

\S4method{\%*\%}{dgeMatrix,dgeMatrix}(x, y)

\S4method{\%*\%}{dgeMatrix,dgCMatrix}(x, y)

\S4method{\%*\%}{dgeMatrix,bis620_sparse_matrix}(x, y)
}
\arguments{
\item{x}{dgeMatrix}

\item{y}{bis620_sparse_matrix}
}
\description{
bis620_matrix_multiply_1

bis620_matrix_multiply_1

bis620_matrix_multiply_1

bis620_matrix_multiply_1

bis620_matrix_multiply_1

bis620_matrix_multiply_1

bis620_matrix_multiply_1

bis620_matrix_multiply_1

bis620_matrix_multiply_1

bis620_matrix_multiply_1

bis620_matrix_multiply_1

bis620_matrix_multiply_1
}
