% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{*,bis620_sparse_matrix,dgeMatrix-method}
\alias{*,bis620_sparse_matrix,dgeMatrix-method}
\alias{*,bis620_sparse_matrix,dgCMatrix-method}
\alias{*,bis620_sparse_matrix,bis620_sparse_matrix-method}
\alias{*,dgCMatrix,dgeMatrix-method}
\alias{*,dgCMatrix,dgCMatrix-method}
\alias{*,dgCMatrix,bis620_sparse_matrix-method}
\alias{*,dgeMatrix,dgeMatrix-method}
\alias{*,dgeMatrix,dgCMatrix-method}
\alias{*,dgeMatrix,bis620_sparse_matrix-method}
\title{bis620_multiply_1}
\usage{
\S4method{*}{bis620_sparse_matrix,dgeMatrix}(e1, e2)

\S4method{*}{bis620_sparse_matrix,dgCMatrix}(e1, e2)

\S4method{*}{bis620_sparse_matrix,bis620_sparse_matrix}(e1, e2)

\S4method{*}{dgCMatrix,dgeMatrix}(e1, e2)

\S4method{*}{dgCMatrix,dgCMatrix}(e1, e2)

\S4method{*}{dgCMatrix,bis620_sparse_matrix}(e1, e2)

\S4method{*}{dgeMatrix,dgeMatrix}(e1, e2)

\S4method{*}{dgeMatrix,dgCMatrix}(e1, e2)

\S4method{*}{dgeMatrix,bis620_sparse_matrix}(e1, e2)
}
\arguments{
\item{e1}{dgeMatrix}

\item{e2}{bis620_sparse_matrix}
}
\description{
bis620_multiply_1

bis620_multiply_1

bis620_multiply_1

bis620_multiply_1

bis620_multiply_1

bis620_multiply_1

bis620_multiply_1

bis620_multiply_1

bis620_multiply_1
}
