% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{bis620_sparse_add}
\alias{bis620_sparse_add}
\title{bis620_sparse_add}
\usage{
bis620_sparse_add(a, b)
}
\arguments{
\item{a}{sparse matrixs or dgeMatrixs or dgCMatrixs or matrixs}

\item{b}{sparse matrixs or dgeMatrixs or dgCMatrixs or matrixs}
}
\value{
a sparse matrix
}
\description{
add two input matrix together
}
