% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{bis620_sparse_divide}
\alias{bis620_sparse_divide}
\title{bis620_divide}
\usage{
bis620_sparse_divide(a, b)
}
\arguments{
\item{a}{sparse matrixs or dgeMatrixs or dgCMatrixs or matrixs}

\item{b}{sparse matrixs or dgeMatrixs or dgCMatrixs or matrixs}
}
\value{
a sparse matrix
}
\description{
the first input matrix divide the second matrix
}
