% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis620_sparse_matrix.R
\name{bis620_sparse_matrix}
\alias{bis620_sparse_matrix}
\title{Creating Sparse Matrix}
\usage{
bis620_sparse_matrix(i, j, x)
}
\arguments{
\item{i}{number of rows}

\item{j}{number of columns}

\item{x}{cell value corresponding to the specified row and column}
}
\value{
min_list columns with the maximum number of NA
}
\description{
This function helps create a sparse matrix which contains only non-zero values cells
}
\examples{

x1 <- bis620_sparse_matrix(
    i = c(1, 2, 5, 6),
    j = c(2, 2, 6, 1),
    x = c(4.3, 5.6, 7, 10)
    )

}
