% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{bis620_sparse_matrix_multiply}
\alias{bis620_sparse_matrix_multiply}
\title{bis620_matrix_multiply}
\usage{
bis620_sparse_matrix_multiply(a, b)
}
\arguments{
\item{a}{sparse matrixs or dgeMatrixs or dgCMatrixs or matrixs}

\item{b}{sparse matrixs or dgeMatrixs or dgCMatrixs or matrixs}
}
\value{
a sparse matrix
}
\description{
matrix multiplication of the two input matrixs
}
