% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{bis620_sparse_minus}
\alias{bis620_sparse_minus}
\title{bis620_minus}
\usage{
bis620_sparse_minus(a, b)
}
\arguments{
\item{a}{sparse matrixs or dgeMatrixs or dgCMatrixs or matrixs}

\item{b}{sparse matrixs or dgeMatrixs or dgCMatrixs or matrixs}
}
\value{
a sparse matrix
}
\description{
the first input matrix minus the second matrix
}
