% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{bis620_sparse_transpose}
\alias{bis620_sparse_transpose}
\title{bis620_sparse_transpose}
\usage{
bis620_sparse_transpose(a)
}
\arguments{
\item{a}{sparse matrix or dgeMatrix or dgCMatrix or matrix}
}
\value{
transpose of the input matrix
}
\description{
get the transpose of the matrix
}
