% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{t,bis620_sparse_matrix-method}
\alias{t,bis620_sparse_matrix-method}
\title{bis620_sparse_transpose}
\usage{
\S4method{t}{bis620_sparse_matrix}(x)
}
\arguments{
\item{x}{sparse matrix}
}
\description{
bis620_sparse_transpose
}
