% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework_function.R
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{Convert Celsius to Kelvin}
\usage{
celsius_to_kelvin(temp_C)
}
\arguments{
\item{temp_C}{a temperature in Celsius}
}
\value{
temperature in Kelvin
}
\description{
This function will convert a temperature from Celsius to Kelvin
}
