% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{Clustering}
\alias{Clustering}
\title{Obtain co-expressed gene modules from co-expression networks}
\usage{
Clustering(adj, num_clusters = 10, min_size = 10)
}
\arguments{
\item{adj}{a symmetric adjacency matrix (co-expression network)}

\item{num_clusters}{the total number of clusters}

\item{min_size}{the minimum size of a cluster to be preserved}
}
\value{
a list of two components, the first component is the cluster assignments of all genes,
the second component is the members in each cluster
}
\description{
Obtain co-expressed gene modules from co-expression networks
}
