% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpress_net.R
\name{CoExpNet}
\alias{CoExpNet}
\title{Estimate co-expression networks with Pearson or Spearman correlation from normalized count data}
\usage{
CoExpNet(n_counts, method = "pearson")
}
\arguments{
\item{n_counts}{normalized count}

\item{method}{name of the method used to estimate co-expression ("pearson" or "spearman")}
}
\value{
a symmetric co-expression network
}
\description{
Estimate co-expression networks with Pearson or Spearman correlation from normalized count data
}
