% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework-3-function.R
\name{dense_to_sparse}
\alias{dense_to_sparse}
\title{Convert dense matrix to data frame}
\usage{
dense_to_sparse(dense)
}
\arguments{
\item{dense}{A dense matrix.}
}
\description{
Convert a dense matrix to a data frame with three columns.
}
