% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_coexpress_net.R
\name{DifferentialNet}
\alias{DifferentialNet}
\title{Estimate differential co-expression networks for two groups from normalized count data
The differential co-expression network of two groups are defined by the difference
of the co-expression network within each group, coupled with a permutation test to
threshold insigificant entries}
\usage{
DifferentialNet(
  n_counts_1,
  n_counts_2,
  method = "pearson",
  rep = 100,
  seed = NULL
)
}
\arguments{
\item{n_counts_1}{normalized count of group 1}

\item{n_counts_2}{normalized count of group 2}

\item{method}{name of the method used to estimate co-expression ("pearson" or "spearman")}

\item{rep}{the number of repetition for permutation test}

\item{seed}{set random seed to ensure reproducibility of the permutation test}
}
\value{
a symmetric differential co-expression network
}
\description{
Estimate differential co-expression networks for two groups from normalized count data
The differential co-expression network of two groups are defined by the difference
of the co-expression network within each group, coupled with a permutation test to
threshold insigificant entries
}
