% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dl}
\alias{dl}
\title{Panitumumab Clinical Trial Data}
\description{
A clinical trial data of a randomized phase 3 study to compare the efficacy
of Panitumumab in combination with FLOFOX to the efficacy of FLOFOX alone in
patients with previously untreated metastatic colorectal cancer.
}
\details{
\code{adae} records the adverse events;
\code{adrsp} records the response;
\code{adsl} records the baseline characteristics, treatment and survival status;
\code{adbiomarker} records the KRAS status;
}
\keyword{data}
