% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichgo.R
\name{EnrichGO}
\alias{EnrichGO}
\title{Obtain GO enrichment terms for a co-expressed gene modules}
\usage{
EnrichGO(clusters, universe = NULL)
}
\arguments{
\item{clusters}{a list of gene clusters (co-expressed gene modules)}

\item{universe}{background genes to correct for}
}
\value{
a list of GO enrichment results for each gene cluster
}
\description{
Obtain GO enrichment terms for a co-expressed gene modules
}
