% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework_function.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Convert Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(temp_F)
}
\arguments{
\item{temp_F}{a temperature in Fahrenheit}
}
\value{
temperature in Celsius
}
\description{
This function will convert a temperature from Fahrenheit to Celsius
}
