% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW1.R
\name{get_baseline}
\alias{get_baseline}
\title{get_baseline}
\usage{
get_baseline(x)
}
\arguments{
\item{x}{a}
}
\value{
NA or a number
}
\description{
This function gets the baseline
}
\examples{
get_baseline(tibble(HGBchange_1=NULL,HGBchange_2=0,HGBchange_3=2,HGBchange_4=NULL))
}
