% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam-utils.R
\name{get_best_resp}
\alias{get_best_resp}
\title{Get the Best Response in ADSL Data}
\usage{
get_best_resp(x)
}
\arguments{
\item{x}{an ADSL dataset for a single patient with RSRESP.}
}
\value{
The best response for the patient.
}
\description{
TODO: add this later
}
