% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_na.R
\name{get_na}
\alias{get_na}
\title{Get list of missing values of the data set}
\usage{
get_na(data)
}
\arguments{
\item{data}{select a data set needed to inspect}
}
\value{
missing_list a list of number of missing values of each column
}
\description{
This function helps examine the number of missing values within factors within a data set
}
\examples{

df <- data.frame(a = c(1,2,3), b = c(NA,0,1), c = c(NA, NA, 0))
df

get_na(df)

}
