% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{get_worst_ae}
\alias{get_worst_ae}
\title{get_worst_ae}
\usage{
get_worst_ae(x)
}
\arguments{
\item{x}{a}
}
\value{
"Mild", "Moderate", "Severe", "Life threatening", "Fatal"
}
\description{
This function gets the worst adverse event through the whole process
}
