% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{GetData}
\alias{GetData}
\title{Get the normalized count data for a cell type and disease status}
\usage{
GetData(
  genes,
  dir = "data-raw//snRNA_seq_PNAS_AD.rds",
  cell_type = "Ex",
  status = "NC"
)
}
\arguments{
\item{genes}{index or names of the genes to be analyzed}

\item{dir}{directory of the dataset}

\item{cell_type}{one of c("Ex", "Mic", "Oli", "Ast", "In")}

\item{status}{one of c("AD", "NC")}
}
\description{
Get the normalized count data for a cell type and disease status
}
