% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello}
\usage{
hello(world = TRUE, verbose = FALSE)
}
\arguments{
\item{world}{a logical indicating if we are saying hello
to the world. (Default TRUE)}

\item{verbose}{should the return be printed? (Default FALSE)}
}
\value{
nothing
}
\description{
This function says hello.
}
\examples{
hello(FALSE)
}
