% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis620_sparse_matrix.R
\name{is.bis620_sparse_matrix}
\alias{is.bis620_sparse_matrix}
\title{is.bis620_sparse_matrix}
\usage{
is.bis620_sparse_matrix(x)
}
\arguments{
\item{x}{bis620_sparse_matrix}
}
\description{
This function is.bis620_sparse_matrix
}
