% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_rank.R
\name{log_rank}
\alias{log_rank}
\title{Run a log_rank test.}
\usage{
log_rank(
  x,
  measure = c("PFS", "OS"),
  arm = c("Treatment", "Rash Grade"),
  filter = c("Wild-type", "Mutant")
)
}
\arguments{
\item{x}{a data.frame records patients' mutant situation, as well as
their baseline characteristics, treatments and survival status.}

\item{measure}{Must be one of "PFS"and "OS". If "PFS", the
function will run log_rank test on Progression-free Survival(PFS). If "OS",
the function will run log_rank test on Overall Survival(OS).}

\item{arm}{Must be one of "Treatment" and "Rash Grade". If "Treatment",
"Treatment" will be predictor of the survival model. If "Rash Grade",
"Rash Grade" will be predictor of the survival model.}

\item{filter}{Must be one of "Wild-type" and "Mutant". If "Wild-type", the
function will run log_rank test for Wild-type KRAS. If "Mutant", the
function will run log_rank test for Mutant KRAS.}
}
\value{
the median(95\% CI) and p-value of the log-rank test.
}
\description{
Based on the given data, run a log_rank test to find out if there is
statistically significant difference.
}
