% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis620_sparse_matrix.R
\name{mat_mult.bis620_sparse_matrix}
\alias{mat_mult.bis620_sparse_matrix}
\title{matrix multiplication of bis620_sparse_matrix}
\usage{
mat_mult.bis620_sparse_matrix(x, y)
}
\arguments{
\item{x}{bis620_sparse_matrix or normal matrix}

\item{y}{bis620_sparse_matrix or normal matrix}
}
\description{
This function performs plus operation of bis620_sparse_matrix
}
