% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework-3-function.R
\name{\%*\%,bis620_sparse_matrix,bis620_sparse_matrix-method}
\alias{\%*\%,bis620_sparse_matrix,bis620_sparse_matrix-method}
\alias{\%*\%,bis620_sparse_matrix,matrix-method}
\alias{\%*\%,bis620_sparse_matrix,dgeMatrix-method}
\alias{\%*\%,bis620_sparse_matrix,dgCMatrix-method}
\alias{\%*\%,dgCMatrix,bis620_sparse_matrix-method}
\title{Matrix multiplication operator performs on two matrix}
\usage{
\S4method{\%*\%}{bis620_sparse_matrix,bis620_sparse_matrix}(x, y)

\S4method{\%*\%}{bis620_sparse_matrix,matrix}(x, y)

\S4method{\%*\%}{bis620_sparse_matrix,dgeMatrix}(x, y)

\S4method{\%*\%}{bis620_sparse_matrix,dgCMatrix}(x, y)

\S4method{\%*\%}{dgCMatrix,bis620_sparse_matrix}(x, y)
}
\arguments{
\item{x}{dgCMatrix}

\item{y}{bis620_sparse_matrix}
}
\description{
Matrix multiplication operator performs on two matrix
}
