% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_BH.R
\name{MatrixBH}
\alias{MatrixBH}
\title{BH Procedure for multiple testing of a matrix}
\usage{
MatrixBH(p_matrix)
}
\arguments{
\item{p_matrix}{a symmetric matrix of p-values}
}
\value{
a symmetric matrix of p-values after BH-procedure
}
\description{
BH Procedure for multiple testing of a matrix
}
