% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW1.R
\name{MEDHXCD_label}
\alias{MEDHXCD_label}
\title{MEDHXCD_label}
\usage{
MEDHXCD_label(x)
}
\arguments{
\item{x}{a number, medical history code}
}
\value{
a disease
}
\description{
This function matches medical history code to specific medical history name
}
\examples{
MEDHXCD_label(2)
}
