% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework-1-function.R
\name{num_na}
\alias{num_na}
\title{Calculate the number of missing values}
\usage{
num_na(x)
}
\arguments{
\item{x}{an R object to be tested: atomic vectors, lists, pairlists, and NULL.}
}
\value{
The sum of the number of the missing values.
}
\description{
Return the sum of all the missing values present in its arguments.
}
