% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam-utils.R
\name{plot_ae}
\alias{plot_ae}
\title{Plot the Number of AE's Longitudinally}
\usage{
plot_ae(x)
}
\arguments{
\item{x}{an AE data set for a single patient with AESTDYI and
AESOC.}
}
\value{
The worst ae for the patient.
}
\description{
TODO: add this later
}
