% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{plot_hgb}
\alias{plot_hgb}
\title{plot the hgb}
\arguments{
\item{x}{a tibble with column "STUDYWEEK" and "value"}
}
\value{
a ggplot object
}
\description{
This function plots the relationship between STUDYWEEK and hgbvalue
}
