% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework-3-function.R
\name{+,bis620_sparse_matrix,bis620_sparse_matrix-method}
\alias{+,bis620_sparse_matrix,bis620_sparse_matrix-method}
\alias{+,bis620_sparse_matrix,matrix-method}
\alias{+,dgeMatrix,bis620_sparse_matrix-method}
\alias{+,bis620_sparse_matrix,dgCMatrix-method}
\title{Plus operator performs on two matrix}
\usage{
\S4method{+}{bis620_sparse_matrix,bis620_sparse_matrix}(e1, e2)

\S4method{+}{bis620_sparse_matrix,matrix}(e1, e2)

\S4method{+}{dgeMatrix,bis620_sparse_matrix}(e1, e2)

\S4method{+}{bis620_sparse_matrix,dgCMatrix}(e1, e2)
}
\arguments{
\item{e1}{bis620_sparse_matrix}

\item{e2}{dgCMatrix}
}
\description{
Plus operator performs on two matrix
}
