% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis620_sparse_matrix.R
\name{print.bis620_sparse_matrix}
\alias{print.bis620_sparse_matrix}
\title{Print bis620_sparse_matrix}
\usage{
\method{print}{bis620_sparse_matrix}(x)
}
\arguments{
\item{x}{bis620_sparse_matrix}
}
\description{
This function prints the layout of bis620_sparse_matrix
}
