% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework-3-function.R
\name{print,bis620_sparse_matrix-method}
\alias{print,bis620_sparse_matrix-method}
\title{Print bis620_sparse_matrix}
\usage{
\S4method{print}{bis620_sparse_matrix}(x)
}
\arguments{
\item{x}{A bis620_sparse_matrix.}
}
\value{
A sparse matrix in general matrix layout.
}
\description{
Print a bis620_sparse_matrix in general matrix layout.
}
