% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis620_sparse_matrix.R
\name{/}
\alias{/}
\title{matrix division of bis620_sparse_matrix}
\usage{
`/`(x, y)
}
\arguments{
\item{x}{bis620_sparse_matrix or normal matrix}

\item{y}{bis620_sparse_matrix or normal matrix}
}
\description{
This function performs subtraction operation of bis620_sparse_matrix
}
