% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis620_sparse_matrix.R
\name{sparse2dense}
\alias{sparse2dense}
\title{sparse2dense}
\usage{
sparse2dense(x)
}
\arguments{
\item{x}{bis620_sparse_matrix}
}
\description{
This function changes sparse matrix into normal matrix
}
