% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework-1-function.R
\name{status}
\alias{status}
\title{Get the larger value of the two values}
\usage{
status(x)
}
\arguments{
\item{x}{a numeric value.}

\item{y}{a numeric value.}
}
\value{
The larger value of the two values.
}
\description{
Return the larger one of the two values.
}
