% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis620_sparse_matrix.R
\name{t,bis620_sparse_matrix-method}
\alias{t,bis620_sparse_matrix-method}
\title{matrix transpose `t()` for bis620_sparse_matrix}
\usage{
\S4method{t}{bis620_sparse_matrix}(x)
}
\description{
matrix transpose `t()` for bis620_sparse_matrix
}
