% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis620_sparse_matrix.R
\name{*,bis620_sparse_matrix,bis620_sparse_matrix-method}
\alias{*,bis620_sparse_matrix,bis620_sparse_matrix-method}
\title{matrix multiplication of bis620_sparse_matrix}
\usage{
\S4method{*}{bis620_sparse_matrix,bis620_sparse_matrix}(e1, e2)
}
\arguments{
\item{x}{bis620_sparse_matrix or normal matrix}

\item{y}{bis620_sparse_matrix or normal matrix}
}
\description{
This function performs subtraction operation of bis620_sparse_matrix
}
