% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework-3-function.R
\name{t,bis620_sparse_matrix-method}
\alias{t,bis620_sparse_matrix-method}
\title{Transpose for bis620_sparse_matrix}
\usage{
\S4method{t}{bis620_sparse_matrix}(x)
}
\arguments{
\item{x}{bis620_sparse_matrix.}
}
\value{
A transposed bis620_sparse_matrix.
}
\description{
Transpose function for bis620_sparse_matrix.
}
