% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_na.R
\name{get_max_na}
\alias{get_max_na}
\title{Get columns with the maximum number of NA}
\usage{
get_max_na(data)
}
\arguments{
\item{data}{select a data set containing columns to get maximum number of NA}
}
\value{
min_list columns with the maximum number of NA
}
\description{
This function helps get the columns with the maximum number of NA within a data set
}
\examples{

df <- data.frame(a = c(1,2,3), b = c(NA,0,1), c = c(NA, NA, 0))
df

get_max_na(df)

}
