% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW1.R, R/adam-utils.R
\name{get_worst_ae}
\alias{get_worst_ae}
\title{get_worst_ae}
\usage{
get_worst_ae(x)

get_worst_ae(x)
}
\arguments{
\item{x}{an AE data set for a single patient with AESEV.}
}
\value{
"Mild", "Moderate", "Severe", "Life threatening", "Fatal"

The worst ae for the patient.
}
\description{
This function gets the worst adverse event through the whole process

TODO: add this later
}
