% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam-utils.R
\name{get_worst_ae}
\alias{get_worst_ae}
\title{Get the Worst AE}
\usage{
get_worst_ae(x)
}
\arguments{
\item{x}{an AE data set for a single patient with AESEV.}
}
\value{
The worst ae for the patient.
}
\description{
TODO: add this later
}
