% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW1.R
\name{plot_hgb}
\alias{plot_hgb}
\title{plot_hgb}
\usage{
plot_hgb(x)
}
\arguments{
\item{x}{a tibble with column "STUDYWEEK" and "value"}
}
\value{
a ggplot object
}
\description{
This function plots the relationship between STUDYWEEK and hgbvalue
}
\examples{
plot_hgb(data.frame(STUDYWEEK=c(1,2,3,4),value=c(3.1,5.2,1.3,8.4))
}
