% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework-1-function.R
\name{status}
\alias{status}
\title{Get the status of a given response}
\usage{
status(x)
}
\arguments{
\item{x}{a character: the response of the patient.}
}
\value{
The status of a given response.
}
\description{
Return the status of a given response.
}
